+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/table-menu/o-table-menu.component.ts
@@ -15,12 +15,10 @@ import {
   ViewEncapsulation
 } from '@angular/core';
 import { MatDialog, MatMenu } from '@angular/material';
-import { BehaviorSubject, Observable, Subscription } from 'rxjs';
-
+import { Observable } from 'rxjs';
 import { InputConverter } from '../../../../../decorators/input-converter';
 import { OTableMenu } from '../../../../../interfaces/o-table-menu.interface';
 import { IReportService } from '../../../../../interfaces/report-on-demand-service.interface';
-
 import { DialogService } from '../../../../../services/dialog.service';
 import { O_REPORT_SERVICE } from '../../../../../services/factories';
 import { SnackBarService } from '../../../../../services/snackbar.service';
@@ -34,14 +32,21 @@ import { OTableCellRendererImageComponent } from '../../../column/cell-renderer/
 import { OColumn } from '../../../column/o-column.class';
 import { OTableComponent } from '../../../o-table.component';
 import { OTableGroupByColumnsDialogComponent } from '../../dialog';
-import { OTableApplyConfigurationDialogComponent } from '../../dialog/apply-configuration/o-table-apply-configuration-dialog.component';
+import {
+  OTableApplyConfigurationDialogComponent
+} from '../../dialog/apply-configuration/o-table-apply-configuration-dialog.component';
 import { OTableExportDialogComponent } from '../../dialog/export/o-table-export-dialog.component';
 import { OTableLoadFilterDialogComponent } from '../../dialog/load-filter/o-table-load-filter-dialog.component';
-import { OTableStoreConfigurationDialogComponent } from '../../dialog/store-configuration/o-table-store-configuration-dialog.component';
+import {
+  OTableStoreConfigurationDialogComponent
+} from '../../dialog/store-configuration/o-table-store-configuration-dialog.component';
 import { OTableStoreFilterDialogComponent } from '../../dialog/store-filter/o-table-store-filter-dialog.component';
 import { OTableVisibleColumnsDialogComponent } from '../../dialog/visible-columns/o-table-visible-columns-dialog.component';
 import { OTableOptionComponent } from '../table-option/o-table-option.component';
 import { OTableExportConfiguration } from './o-table-export-configuration.class';
+
+
+
 
 export const DEFAULT_INPUTS_O_TABLE_MENU = [
   // select-all-checkbox [yes|no|true|false]: show selection check boxes. Default: no.
@@ -61,8 +66,9 @@ export const DEFAULT_INPUTS_O_TABLE_MENU = [
 
   // show-group-by-option [yes|no|true|false]: show group by menu option in the header menu
   'showGroupByOption: show-group-by-option',
-
-   // show-report-on-demand-option [yes|no|true|false]: show report on demand option in the header menu
+  // show-reset-width-option [yes|no|true|false]: show reset width menu option in the header menu
+  'showResetWidthOption: show-reset-width-option',
+  // show-report-on-demand-option [yes|no|true|false]: show report on demand option in the header menu
   'showReportOnDemandOption: show-report-on-demand-option'
 ];
 
@@ -96,7 +102,10 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
   @InputConverter()
   showGroupByOption: boolean = true;
   @InputConverter()
+  showResetWidthOption: boolean = true;
+  @InputConverter()
   showReportOnDemandOption: boolean = true;
+
 
   public onVisibleFilterOptionChange: EventEmitter<any> = new EventEmitter();
   /* End of inputs */
@@ -125,13 +134,9 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
   @ViewChild('columnFilterOption', { static: false })
   columnFilterOption: OTableOptionComponent;
 
-  private showColumnsFilterOptionSubject = new BehaviorSubject<boolean>(false);
-  public showColumnsFilterOption: Observable<boolean> = this.showColumnsFilterOptionSubject.asObservable();
-
   protected permissions: OTableMenuPermissions;
   protected mutationObservers: MutationObserver[] = [];
 
-  private subscription: Subscription;
 
   constructor(
     protected injector: Injector,
@@ -143,9 +148,6 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
     this.dialogService = this.injector.get(DialogService);
     this.translateService = this.injector.get(OTranslateService);
     this.snackBarService = this.injector.get(SnackBarService);
-    const self = this;
-
-    this.subscription = this.onVisibleFilterOptionChange.subscribe((x: boolean) => self.showColumnsFilterOptionSubject.next(x));
   }
 
   ngOnInit() {
@@ -157,8 +159,6 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
   }
 
   ngAfterViewInit() {
-
-    this.showColumnsFilterOptionSubject.next(this.table.oTableColumnsFilterComponent !== undefined);
 
     if (!this.permissions.items || this.permissions.items.length === 0) {
       return;
@@ -201,7 +201,6 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
         m.disconnect();
       });
     }
-    this.subscription.unsubscribe();
   }
 
   registerOptions(oTableOptions: OTableOptionComponent[]) {
@@ -477,6 +476,14 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
       if (result) {
         this.table.clearFilters();
         this.table.reloadPaginatedDataFromStart(false);
+      }
+    });
+  }
+
+  onResetWidthClicked() {
+    this.dialogService.confirm('CONFIRM', 'TABLE.DIALOG.CONFIRM_RESET_WIDTH').then(result => {
+      if (result) {
+        this.table.resetColumnsWidth();
       }
     });
   }