+++ projects/ontimize-web-ngx/src/lib/components/table/o-table.component.ts
@@ -48,6 +48,7 @@ import { ComponentStateServiceProvider, O_COMPONENT_STATE_SERVICE, OntimizeServi
 import { SnackBarService } from '../../services/snackbar.service';
 import { OTableComponentStateClass } from '../../services/state/o-table-component-state.class';
 import { OTableComponentStateService } from '../../services/state/o-table-component-state.service';
+import { OColumnDisplay } from '../../types';
 import { Expression } from '../../types/expression.type';
 import { OPermissions } from '../../types/o-permissions.type';
 import { OQueryDataArgs } from '../../types/query-data-args.type';
@@ -224,7 +225,10 @@ export const DEFAULT_INPUTS_O_TABLE = [
   'showExpandableIconFunction: show-expandable-icon-function',
 
   // show-report-on-demand-option [yes|no|true|false]: show report on demand option in the table menu. Default: yes.
-  'showReportOnDemandOption: show-report-on-demand-option'
+  'showReportOnDemandOption: show-report-on-demand-option',
+
+  // show-reset-width-option [yes|no|true|false]: show reset width menu option in the header menu
+  'showResetWidthOption: show-reset-width-option'
 ];
 
 export const DEFAULT_OUTPUTS_O_TABLE = [
@@ -353,6 +357,8 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
   showButtonsText: boolean = true;
   @InputConverter()
   filterColumnActiveByDefault: boolean = true;
+  @InputConverter()
+  showResetWidthOption: boolean = true;
 
   // Expandable input callback function
   showExpandableIconFunction: (row: any, rowIndex: number) => boolean | Promise<boolean> | Observable<boolean>;
@@ -461,6 +467,7 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
   public daoTable: OTableDao | null;
   public dataSource: OTableDataSource | null;
   public visibleColumns: string;
+  public searcheableColumns: string[] = [];
   public defaultVisibleColumns: string;
   public groupedColumns: string;
 
@@ -878,8 +885,11 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
         if (!this.oTableColumnsFilterComponent) {
           this.oTableColumnsFilterComponent = new OTableColumnsFilterComponent(this.injector, this);
           this.oTableMenu.onVisibleFilterOptionChange.next(this.filterColumnActiveByDefault);
+          this.oTableColumnsFilterComponent.columns = clonedOpts.filterColumns;
         }
-        this.oTableColumnsFilterComponent.columns = clonedOpts.filterColumns;
+        else {
+          this.oTableColumnsFilterComponent.columns = this.searcheableColumns.join(";");
+        }
       }
     }
 
@@ -892,6 +902,7 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
 
   protected initTableAfterViewInit() {
     this.parseVisibleColumns();
+    this.parseSearcheableColumns();
     this.setDatasource();
     this.registerDataSourceListeners();
     this.parseGroupedColumns();
@@ -906,7 +917,13 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
       this.queryData();
     }
   }
+  parseSearcheableColumns() {
+    this.searcheableColumns = this.visibleColArray.filter(col => {
+      const oCol = this.getOColumn(col);
+      return oCol && oCol.searchable;
+    })
 
+  }
   destroy() {
     super.destroy();
     if (this.tabGroupChangeSubscription) {
@@ -1010,19 +1027,18 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
       // a o-table-column definition trying to replace an already existing o-table-column definition
       return;
     }
+
     const colDef: OColumn = this.createOColumn(column.attr, this, column);
     let columnWidth = column.width;
-
     const storedData = this.state.getColumnDisplay(colDef);
     if (Util.isDefined(storedData) && Util.isDefined(storedData.width)) {
       // check that the width of the columns saved in the initial configuration
       // in the local storage is different from the original value
       if (this.state.initialConfiguration.columnsDisplay) {
-        this.state.initialConfiguration.columnsDisplay.forEach(element => {
-          if (colDef.attr === element.attr && element.width === colDef.definition.originalWidth) {
-            columnWidth = storedData.width;
-          }
-        });
+        const initialStoredData = this.state.initialConfiguration.getColumnDisplay(colDef);
+        if (initialStoredData && initialStoredData.width === colDef.definition.originalWidth) {
+          columnWidth = storedData.width;
+        }
       } else {
         columnWidth = storedData.width;
       }
@@ -1076,10 +1092,10 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
     }
   }
 
-  parseVisibleColumns() {
+  parseVisibleColumns(defaultConfiguration: boolean = false) {
     if (this.state.columnsDisplay) {
       // filtering columns that might be in state storage but not in the actual table definition
-      let stateCols = [];
+      let stateCols: OColumnDisplay[] = [];
       this.state.columnsDisplay.forEach((oCol, index) => {
         const isVisibleColInColumns = this._oTableOptions.columns.find(col => col.attr === oCol.attr) !== undefined;
         if (isVisibleColInColumns) {
@@ -1088,15 +1104,20 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
           console.warn('Unable to load the column ' + oCol.attr + ' from the localstorage');
         }
       });
-      stateCols = this.checkChangesVisibleColummnsInInitialConfiguration(stateCols);
+      if (defaultConfiguration) {
+        stateCols = this.state.initialConfiguration.columnsDisplay;
+      } else {
+        stateCols = this.checkChangesVisibleColummnsInInitialConfiguration(stateCols);
+      }
       this.visibleColArray = stateCols.filter(item => item.visible).map(item => item.attr);
+
     } else {
       this.visibleColArray = Util.parseArray(this.defaultVisibleColumns ? this.defaultVisibleColumns : this.visibleColumns, true);
       this._oTableOptions.columns.sort((a: OColumn, b: OColumn) => this.visibleColArray.indexOf(a.attr) - this.visibleColArray.indexOf(b.attr));
     }
   }
 
-  checkChangesVisibleColummnsInInitialConfiguration(stateCols) {
+  checkChangesVisibleColummnsInInitialConfiguration(stateCols: OColumnDisplay[]): OColumnDisplay[] {
     if (this.state.initialConfiguration.columnsDisplay) {
       const originalVisibleColArray =
         this.state.initialConfiguration.columnsDisplay.filter(x => x.visible).map(x => x.attr);
@@ -1141,7 +1162,30 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
     }
     return stateCols;
   }
+  changeColumnsVisibility(colToAddInVisibleCol, stateCols) {
+    colToAddInVisibleCol.forEach((colAdd, index) => {
+      if (stateCols.filter(col => col.attr === colAdd).length > 0) {
+        stateCols = stateCols.map(col => {
+          if (colToAddInVisibleCol.indexOf(col.attr) > -1) {
+            col.visible = true;
+          }
+          return col;
+        });
+      } else {
+        this.colArray.forEach((element, i) => {
+          if (element === colAdd) {
+            stateCols.splice(i + 1, 0,
+              {
+                attr: colAdd,
+                visible: true,
+                width: undefined
+              });
+          }
 
+        });
+      }
+    });
+  }
   parseSortColumns() {
     const sortColumnsParam = this.state.sortColumns || this.sortColumns;
     this.sortColArray = ServiceUtils.parseSortColumns(sortColumnsParam);
@@ -1180,6 +1224,7 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
         this.sortColArray.splice(i, 1);
       }
     }
+
   }
 
   protected ensureColumnsOrder() {
@@ -2148,10 +2193,11 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
   }
 
   /**
-   * Clear all filters(column filter and quickfilter) and reload data
+   * Clear all filters(column filter, quickfilter and filter builder) and reload data
    * @param [triggerDatasourceUpdate]
    */
   clearFilters(triggerDatasourceUpdate: boolean = true): void {
+
     this.dataSource.clearColumnFilters(triggerDatasourceUpdate);
     if (this.oTableMenu && this.oTableMenu.columnFilterOption) {
       this.oTableMenu.columnFilterOption.setActive(this.isColumnFiltersActive);
@@ -2160,6 +2206,12 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
     if (this.oTableQuickFilterComponent) {
       this.oTableQuickFilterComponent.setValue(void 0);
     }
+
+    if (this.filterBuilder) {
+      this.filterBuilder.clearFilter();
+      this.filterBuilder.triggerReload();
+    }
+
   }
 
   clearColumnFilter(attr: string, triggerDatasourceUpdate: boolean = true): void {
@@ -2181,9 +2233,11 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
   }
 
   isColumnFilterable(column: OColumn): boolean {
-    return (this.oTableColumnsFilterComponent && this.oTableColumnsFilterComponent.isColumnFilterable(column.attr));
+    return Util.isDefined(this.oTableColumnsFilterComponent) ? this.oTableColumnsFilterComponent.isColumnFilterable(column.attr) : this.isSearcheableColumn(column);
   }
-
+  isSearcheableColumn(column: OColumn): boolean {
+    return this.searcheableColumns.includes(column.attr);
+  }
   isColumnFilterActive(column: OColumn): boolean {
     return this.isColumnFiltersActive && Util.isDefined(this.dataSource.getColumnValueFilterByAttr(column.attr));
   }
@@ -2197,8 +2251,8 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
         column: column,
         activeSortDirection: this.getSortFilterColumn(column),
         tableData: this.dataSource.getCurrentData(),
-        preloadValues: this.oTableColumnsFilterComponent.preloadValues,
-        mode: this.oTableColumnsFilterComponent.mode,
+        preloadValues: this.oTableColumnsFilterComponent ? this.oTableColumnsFilterComponent.preloadValues : true,
+        mode: this.oTableColumnsFilterComponent ? this.oTableColumnsFilterComponent.mode : 'default',
         startView: this.getStartViewFilterColumn(column)
       },
       minWidth: '380px',
@@ -2551,6 +2605,11 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
         }
       });
     }
+
+    if (this.filterBuilder && Util.isDefined(storage.filterBuilderValues)) {
+      this.filterBuilder.setFilterValues(storage.filterBuilderValues);
+      this.filterBuilder.triggerReload();
+    }
   }
 
   onStoreConfigurationClicked() {
@@ -2567,14 +2626,21 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
 
   applyDefaultConfiguration() {
     this.initializeParams();
-    this.parseVisibleColumns();
+    this.parseVisibleColumns(true);
+    this.reinitializateQuickFilterColumns();
     this._oTableOptions.columns.sort((a: OColumn, b: OColumn) => this.visibleColArray.indexOf(a.attr) - this.visibleColArray.indexOf(b.attr));
-    this.reinitializeSortColumns();
+    this.resetQueryRows();
+    const initialConfigSortColumnsArray = ServiceUtils.parseSortColumns(this.state.initialConfiguration.sortColumns);
+    this.reinitializeSortColumns(initialConfigSortColumnsArray);
     this.onReinitialize.emit(null);
     this.clearFilters(false);
     this.reloadData();
   }
-
+  resetQueryRows() {
+    if (Util.isDefined(this.state.initialConfiguration.queryRows)) {
+      this.queryRows = this.state.initialConfiguration.queryRows;
+    }
+  }
   applyConfiguration(configurationName: string) {
     const storedConfiguration = this.state.getStoredConfiguration(configurationName);
     if (storedConfiguration) {
@@ -2591,6 +2657,7 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
             break;
           case 'quick-filter':
           case 'columns-filter':
+          case 'filter-builder':
             this.setFiltersConfiguration();
             break;
           case 'grouped-columns':
@@ -2950,6 +3017,12 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
     this.componentStateService.storeFilter(arg);
   }
 
+  reinitializateQuickFilterColumns() {
+    this._oTableOptions.columns
+      .filter(column => column.searchable && column.visible)
+      .forEach(column => column.searching = true)
+  }
+
   reinitializeSortColumns(sortColumns?: SQLOrder[]) {
     if (Util.isDefined(sortColumns)) {
       this.sortColArray = sortColumns;
@@ -2982,13 +3055,15 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
     return of(null);
   }
 
-  tableQuickFilterChanged() {
+  tableQuickFilterChanged(value: string) {
     if (this.pageable) {
       this.queryCellRenderers().subscribe(() => {
         this.reloadPaginatedDataFromStart(false);
       });
     } else {
-      this.reloadData(false);
+      if (value && value.length) {
+        this.dataSource.quickFilter = value;
+      }
     }
   }
 
@@ -3037,4 +3112,20 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
     return (Util.isDefined(this.showExpandableIconFunction) && this.showExpandableIconFunction instanceof Function) ? Util.wrapIntoObservable(this.showExpandableIconFunction(row, rowIndex)) : of(true);
   }
 
+  resetColumnsWidth() {
+    this._oTableOptions.columns.forEach(c => {
+      if (Util.isDefined(c.definition)) {
+        c.width = Util.isDefined(c.definition.width) ? c.definition.width : c.definition.originalWidth
+      } else {
+        c.width = "auto";
+      }
+    });
+    this.cd.detectChanges();
+    this.updateColumnsDOMWidth();
+  }
+  updateColumnsDOMWidth() {
+    this._oTableOptions.columns.forEach(c => {
+      c.DOMWidth = this.getThWidthFromOColumn(c);
+    });
+  }
 }