+++ projects/ontimize-web-ngx/src/lib/directives/index.ts
@@ -1,2 +1,3 @@
 export * from './keyboard-listener.directive';
 export * from './o-hidden.directive';
+export * from './o-mat-error.directive';