+++ projects/ontimize-web-ngx/src/lib/types/configure-service-args.type.ts
@@ -1,0 +1,9 @@
+import { Injector } from "@angular/core";
+
+export type OConfigureServiceArgs = {
+  injector:Injector;
+  baseService: any;
+  entity: string;
+  service: string;
+  serviceType: string;
+};