+++ CHANGELOG.md
@@ -1,4 +1,49 @@
 
+## 8.8.0
+### Features:
+* **o-table**:
+  * `Save/load/clear filter` and `Save/load configuration` operations have in consider the filter builder component([b47963b](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/b47963b)) ([209b297](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/209b297)) Closes [#994](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/994)
+  * Modified the order of the table menu options ([54dad824](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/54dad824))
+  * Added a new menu option to reset the columns width ([f5d1f74a](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/f5d1f74a)) Closes [#799](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/799)
+* **o-grid,o-list**: Moved `quick-filter-appearance` to AbstractOServiceComponent ([0bc6c49](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/0bc6c49))
+
+### Bug fixes
+* **o-real-input**: New attribute `strict` ([e466768f](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/e466768f)) ([e264a54](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/e264a54)) Closes [#1022](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1022)
+* **o-table**:
+  * Fixed bug that incorrectly set the default configuration ([1808d599](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/1808d599)) Closes [#1006](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1006)
+  * Fixed bug that prevents saving filters and settings the first time ([207f0856](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/207f0856)) Closes [#1003](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1003)
+  * Fixed bug in `AbstractOServiceBaseComponent` when searching for a column of type service render ([47a60682](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/47a60682)) Closes[#1010](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1010)
+  * `o-table-columns-filter`: active by default. However, you can configure it unavailable in table header with `filter-column-active-by-default= 'no'`. ([6ceaefe5](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/6ceaefe5)) Closes [#1007](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1007)
+  * Fix bug that filtering a value in the quick filter launchs the query to the backend when `pageable=no` ([367f3d7](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/367f3d7)) Closes [#1011](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1011)
+* **o-email-input**: Fixed bug that incorrectly validated email inputs ([46c6e78](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/46c6e78)) Closes [#999](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/999)
+* **mat-error**: Bug [#1013](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1013) fixed. OntimizeWeb form component `mat-error` definition no longer interferes the Angular Material default definition.
+  *NOTES*:
+    * `OMatErrorModule` no longer exists. This shouldn't affect users because its importation wasn't needed.
+    * There is a new `oMatError` directive that user must use instead `ngIf` for having the `O_MAT_ERROR_OPTIONS` features available in the Angular Material `mat-error` component. See documentation for more info.
+* **service-type**: Fixes bug when multiple components use the same service-type ([380fda3](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/380fda3)) Closes [#1015](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1015)
+* **o-hour-input**: Fixing wrong `value-type="timestamp"` value saving ([f3d6f39c](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/f3d6f39c)) ([5b9abf17](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/5b9abf17)) Closes [#960](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/960)
+
+### BREAKING CHANGES
+* Due to the fix of the issue [#1015](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1015), to set the `service-type` attribute on the component, the provider defined in the module cannot set `useFactory`, instead use `useValue` because Ontimize Web is responsible for creating each instance for this service.
+
+* **Application configuration file `app.config,ts`**
+
+The default value of `serviceType` has been changed to `OntimizeEE` and the default value of `permissionsServiceType` has been changed to `OntimizeEEPermissions`([25cc5c8](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/25cc5c8))  Closes [#967](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/967)
+
+
+```ts
+@NgModule({
+
+  ...
+
+  providers: [{
+    provide: 'starWars',
+    useValue: StarWarsService
+  }]
+})
+export class MyModule { }
+```
+
 ## 8.7.3 (2022-10-06)
 ### Features
 * JWT token refresh has been implemented to be compatible with [Ontimize Boot](https://ontimize.github.io/ontimize-boot/basics/autoconfigurators/#security) ([280a422](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/280a422)) Closes [#1041](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1041)
@@ -22,7 +67,6 @@ To resolve this breaking change, you can use the new `O_TABLE_GLOBAL_CONFIG` Inj
   ],
   ...
 })
-```
 
 ## 8.7.2 (2022-09-15)
 ### Bug fixes